requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-google-genai

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Machine Learning
sentence-transformers>=2.2.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0

# OpenAI
openai>=1.0.0

# Web UI
streamlit>=1.28.0

# Additional utilities
typing-extensions>=4.0.0
